requirements.txt
Flask>=3.0,<4.0
RPi.GPIO>=0.7.1
